pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pptx-research-report"
version = "0.1.0"
description = "PPTX-first research report generation pipeline"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2.7,<3.0",
  "pydantic-settings>=2.2,<3.0",
  "python-pptx>=0.6.23,<1.0",
  "typer>=0.12,<1.0",
  "rich>=13.7,<14.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2,<9.0",
  "ruff>=0.4,<1.0",
]

[project.scripts]
reportgen = "reportgen.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
